requirements.txt
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
idna==3.8
outcome==1.3.0.post0
peewee==3.17.6
pillow==10.4.0
PySocks==1.7.1
requests==2.32.3
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
